Add serilization-deserialization mixin
